Now it calculates the speed and writes it to unity_speed.txt, not working with unity so far =(

+++ device/Device.cpp
@@ -10,7 +10,7 @@
 using namespace std;
 
 int Device::pythonPid;
-int Device::buffer[BUFFER_SIZE];
+double Device::buffer[BUFFER_SIZE];
 
 /**
  *	Initialize python program, because we need to give it our pid
